+++ src/Mcamara/LaravelLocalization/LaravelLocalization.php
@@ -97,13 +97,6 @@ class LaravelLocalization {
     protected $routeName;
 
     /**
-     * Name of the locale variable for the session and cookie storages
-     *
-     * @var string
-     */
-    protected $cookieSessionName = 'locale';
-
-    /**
      * Creates new instance.
      * @throws UnsupportedLocaleException
      */
@@ -111,10 +104,10 @@ class LaravelLocalization {
     {
         $this->app = app();
 
-        $this->configRepository = $this->app['config'];
-        $this->view = $this->app['view'];
-        $this->translator = $this->app['translator'];
-        $this->router = $this->app['router'];
+        $this->configRepository = $this->app[ 'config' ];
+        $this->view = $this->app[ 'view' ];
+        $this->translator = $this->app[ 'translator' ];
+        $this->router = $this->app[ 'router' ];
         $this->request = $this->app[ 'request' ];
 
         // set default locale
@@ -150,7 +143,7 @@ class LaravelLocalization {
         {
             // if the first segment/locale passed is not valid
             // the system would ask which locale have to take
-            // it could be taken by session, browser or app default
+            // it could be taken by the browser
             // depending on your configuration
 
             $locale = null;
@@ -162,7 +155,7 @@ class LaravelLocalization {
                 $this->currentLocale = $this->defaultLocale;
             }
             // but if hideDefaultLocaleInURL is false, we have
-            // to retrieve it from the session/cookie/browser...
+            // to retrieve it from the browser...
             else
             {
                 $this->currentLocale = $this->getCurrentLocale();
@@ -170,10 +163,6 @@ class LaravelLocalization {
         }
 
         $this->app->setLocale($this->currentLocale);
-
-        $this->storeSession($this->currentLocale);
-
-        $this->storeCookie($this->currentLocale);
 
         return $locale;
     }
@@ -340,7 +329,6 @@ class LaravelLocalization {
         {
             $route = '/' . $locale;
         }
-
         if ( is_string($locale) && $this->translator->has($transKeyName, $locale) )
         {
             $translation = $this->translator->trans($transKeyName, [ ], "", $locale);
@@ -479,19 +467,6 @@ class LaravelLocalization {
             return $this->currentLocale;
         }
 
-        // get session language...
-        if ( $locale = $this->getSessionLocale() )
-        {
-            return $locale;
-        }
-
-        // or get cookie language...
-        if ( $locale = $this->getCookieLocale() )
-        {
-            return $locale;
-        }
-
-        // or get browser language...
         if ( $this->useAcceptLanguageHeader() )
         {
             $negotiator = new LanguageNegotiator($this->defaultLocale, $this->getSupportedLocales(), $this->request);
@@ -513,60 +488,6 @@ class LaravelLocalization {
         return array_keys($this->supportedLocales);
     }
 
-
-    /**
-     * Returns the locale stored on the session (if available)
-     *
-     * @return string|false Locale stored in session - False if it doesn't exist
-     */
-    protected function getSessionLocale()
-    {
-        if ( $this->useSessionLocale() && app('session')->has($this->cookieSessionName) )
-        {
-            return session($this->cookieSessionName);
-        }
-
-        return false;
-    }
-
-    /**
-     * Store the locale on a session variable
-     * @param $locale string Locale to save on session variable
-     */
-    protected function storeSession( $locale )
-    {
-        if ( $this->useSessionLocale() )
-        {
-            session([ $this->cookieSessionName => $locale ]);
-        }
-    }
-
-    /**
-     * Returns the locale stored on the cookies (if available)
-     *
-     * @return string|false Locale stored in cookies - False if it doesn't exist
-     */
-    protected function getCookieLocale()
-    {
-        if ( $this->useCookieLocale() )
-        {
-            return cookie($this->cookieSessionName)->getValue();
-        }
-
-        return false;
-    }
-
-    /**
-     * Store the locale on a cookie variable
-     * @param $locale String Locale to store on Cookie variable
-     */
-    protected function storeCookie( $locale )
-    {
-        if ( $this->useCookieLocale() )
-        {
-            cookie($this->cookieSessionName, $locale);
-        }
-    }
 
     /**
      * Check if Locale exists on the supported locales array
@@ -639,6 +560,7 @@ class LaravelLocalization {
         {
             $this->translatedRoutes[ ] = $routeName;
         }
+
         return $this->translator->trans($routeName);
     }
 
@@ -663,11 +585,12 @@ class LaravelLocalization {
 
         foreach ( $this->translatedRoutes as $route )
         {
-            if ( $this->substituteAttributesInRoute( $attributes, $this->translator->trans($route)) === $path )
+            if ( $this->substituteAttributesInRoute($attributes, $this->translator->trans($route)) === $path )
             {
                 return $route;
             }
         }
+
         return false;
     }
 
@@ -751,25 +674,6 @@ class LaravelLocalization {
         return $this->configRepository;
     }
 
-    /**
-     * Returns the translation key for a given path
-     *
-     * @return boolean       Returns value of useSessionLocale in config.
-     */
-    protected function useSessionLocale()
-    {
-        return $this->configRepository->get('laravellocalization.useSessionLocale');
-    }
-
-    /**
-     * Returns the translation key for a given path
-     *
-     * @return boolean       Returns value of useCookieLocale in config.
-     */
-    protected function useCookieLocale()
-    {
-        return $this->configRepository->get('laravellocalization.useCookieLocale');
-    }
 
     /**
      * Returns the translation key for a given path
@@ -799,12 +703,14 @@ class LaravelLocalization {
      */
     public function createUrlFromUri( $uri )
     {
+        $uri = ltrim($uri, "/");
+
         if ( empty( $this->baseUrl ) )
         {
             return app('url')->to($uri);
         }
 
-        return $this->baseUrl . ltrim($uri, "/");
+        return $this->baseUrl . $uri;
     }
 
     /**
@@ -904,8 +810,7 @@ class LaravelLocalization {
                 }
             }
 
-        }
-        else
+        } else
         {
             if ( !$this->router->current() )
             {
@@ -925,7 +830,6 @@ class LaravelLocalization {
                 $attributes = array_merge($attributes, $response);
             }
         }
-
 
 
         return $attributes;